tests: nested multiprocessing: do not assume order of results
As per "Pipes and Queues" section of the `multiprocessing`
documentation, when multiple processes are enqueing the objects,
the other end might receive them out-of-order.

See: https://docs.python.org/3/library/multiprocessing.html#pipes-and-queues

Have the `test_multiprocessing_nested_process` accept both
`[1, 2]` and `[2, 1]` as a valid result. This should fix sporadic
failures of the afore-mentioned test when enqueued values happen
to be received out-of-order.