_pyi_rth_utils: fix path matching in prepend_path_to_environment_variable
Fix path matching in `_pyi_rth_utils.prepend_path_to_environment_variable`
when trying to determine if the given path is already stored in
the target environment variable. Instead of checking for occurrence
of the given path within the environment variable's value, we need
to split the value on `os.pathsep` and search the resulting list of
paths. This ensures that we match only the exact (full) path, instead
of erroneously matching a prefix of a longer path (i.e., to a
sub-directory of the given path) that already happens to be among
the stored paths.

This fixes the problem with Qt run-time hooks failing to add the
top-level application directory to `PATH` (necessary for `QtNetwork`
to discover the bundled OpenSSL DLLs) when the `pywin32` run-time
hook alreayd added the path to bundled `pywin32_system32` directory
to `PATH`.