tests: have signal forwarding tests print elapsed time
Have the test application for onefile signal forwarding test print
elapsed time when it successfully receives the signal.

This way, when we have a sporadic test failure, the successful parts
of the test provide indication about how close they were to failing
themselves, and whether the failed run was due to timeout being too
strict or due to some other issue.