bootloader: have forwarding signal handler store/restore errno
Have the forwarding signal handler store the value of errno at the
start, and restore it at the end. At least in theory, the `kill()`
call that forwards the signal could fail, and modify the errno,
which might lead to inconsitent state in the main code once the
signal handler exits.