bootloader: fix race condition with signal handler initialization
In the POSIX codepath, use a semaphore to ensure that the child
process of onefile application enters execution of the python
code only *after* the parent process has fully completed the `fork()`
call *and* stored the child process to `pyi_ctx->child_pid` to
activate the forwarding signal handler.

Without explicit synchronization, it might happen that under heavy
CPU load, the parent process is stalled, while the child process
continues with execution while the signal handlers in the parent
are still no-op, which leads to sporadic failures in our
`test_onefile_signal_handling`.

The SysV semaphore API is used because it is most portable (and
specifically, because un-named POSIX semaphores are not available
on macOS).