bootloader: clean up some earlier changes and simplify comments
Due to added explicit sync between parent and child process via
semaphore in the POSIX codepath, we can move the signal handler
installation back after the `fork()` call. This effectively
reverts 235f6b9, which moved the signal handler installation
before `fork()` in an attempt to mitigate the race condition,
which now turned out to be solvable only via explicit sync.
This way, we can remove the comment block explaining why the
handlers were installed before the `fork()`.

Update the comment under the child-process codepath block; the
part about cleanup being an exception is not relevant anymore,
because we force the child process to exit on `execvp()` failure,
to avoid both processes entering the cleanup codepath and
interfere with each other.