loader: have PyiFrozenLoader's methods validate passed module name
Have the methods of PyiFrozenLoader that accept the module name
(for example, `get_filename`, `get_code`, `get_resource_reader`,
which all have the `fullname` argument) validate the passed name
and raise an ˙ImportError` if the name does not match the name of
module for which the loader instance was created.

This matches the behavior of python's built-in loaders; it also
opens door for further cleanup/simplification of our loader's
methods.