bootloader: allow creation of sync semaphore to fail
Allow creation of sync semaphore in the POSIX onefile codepath to
fail. The synchronization between onefile parent and child process
was introduced in #8900 to fix race condition that was causing
sporadic failures of `test_onefile_signal_handling`.

However, outside of our POSIX signal test suite, the absence of
synchronization should be of little consequence. Therefore, if
we fail to create the semaphore, let the frozen application
continue without synchronization. This gives us pre-#8900 behavior,
which should be preferable to aplication not running at all.

For example, under cygwin, `semget()` is only available when
cygserver is running, which might not always be the case. There
might be other failure modes on other POSIX systems.