hookutils: setuptools: fix errors with de-vendored setuptools
Some linux distributions (notably, Arch linux) completely de-vendor
their `setuptools` package; i.e., install the otherwise vendored
dependencies as external packages, and remove the `setuptools/_vendor`
directory.

It seems de-vendoring was performed even before v71.0 of `setuptools`
(which made the process easier), and it was mostly compatible with
our hooks, because `collect_submodules` gracefully handles the case
when the specified package does not exist. However, when implementing
support for `setuptools` >= 71.0 in #8720, we added an unguarded
attempt at importing `setuptools._vendor`. In case of de-vendored
`setuptools`, this raises a `ModuleNotFoundError`, and prevents
proper initialization of `SetuptoolsInfo`, which results in errors
in hooks that make use of it.

Therefore, add try/except block to gracefully handle module-not-found
error. If `setuptools._vendor` does not exist, none of the packages
are vendored (and we can skip the corresponding code block). Similarly,
we can skip attempts at collecting submodules of `setuptools_vendor`
as well as the select data files; this is mostly to avoid having
`collect_submodules` and `collect_data_files` print warnings/debug
messages about missing package.